Changes have been made to the Education Program completion and Badge earning processes.

+++ Business/Dtos/Requests/AccountEducationProgramRequests/CreateAccountEducationProgramRequest.cs
@@ -4,5 +4,5 @@ public class CreateAccountEducationProgramRequest
 {
     public Guid AccountId { get; set; }
     public Guid EducationProgramId { get; set; }
-    public int StatusPercent { get; set; }
+    public double StatusPercent { get; set; }
 }
